refactor: remove user data from access token response
